Column name for datasets

A string list containing the column name for datasets.

Usage

dataset_info

Format

A string list with 14 entries.

dataset_colname

The string name of the dataset column, "Dataset".

valid_from_version_colname

The string name of the column that gives the initial version for which this datapoint is valid, "ValidFromVersion".

valid_to_version_colname

The string name of the column that gives the last version for which this datapoint is valid, "ValidToVersion".

iea

A string identifying that ECC data are from the IEA exclusively.

mw

A string identifying that ECC data are for muscle work (MW) exclusively.

both

A string identifying that ECC data include both IEA and muscle work.

Examples

dataset_info
#> $dataset_colname
#> [1] "Dataset"
#> 
#> $valid_from_version_colname
#> [1] "ValidFromVersion"
#> 
#> $valid_to_version_colname
#> [1] "ValidToVersion"
#> 
#> $clpfu_iea
#> [1] "CL-PFU IEA"
#> 
#> $clpfu_mw
#> [1] "CL-PFU MW"
#> 
#> $clpfu_iea_mw
#> [1] "CL-PFU IEA+MW"
#> 
#> $clpfu
#> [1] "CL-PFU"
#> 
#> $ieaeweb
#> [1] "IEA EWEB"
#> 
#> $faostat
#> [1] "FAOSTAT"
#> 
#> $ilostat
#> [1] "ILOSTAT"
#> 
#> $wlrpfu
#> [1] "WLR-PFU"
#> 
#> $wlrpfu_elect
#> [1] "WLR-PFU Electricity"
#> 
#> $wlrpfu_trans
#> [1] "WLR-PFU Transport"
#> 
#> $wlrpfu_mw
#> [1] "WLR-PFU Muscle work"
#>